+++ lib/products_feature/remote_products_loader_errors.dart
@@ -1,0 +1,1 @@
+enum RemoteProductsLoaderErrors { connectivity, invalidData }
